\name{biom1}
\alias{biom1}
\docType{data}

\title{
Association between the gut microbiome and colorectal cancer
}

\description{
Real data on colorectal cancer
}

\usage{data(biom1)}
\format{
phyloseq
}

\references{
McMurdie PJ and Holmes S. phyloseq: An R package for reproducible interactive analysis and graphics of microbiome census data. PLoS ONE. 2013;8(4):e61217.

Zackular JP, et al. The human gut microbiome as a screening tool for colorectal cancer. Cancer Prevention Research 2014, 7(11):1112–1121.
}

\examples{
library(phyloseq)
data(biom1)
}
