\name{biom2}
\alias{biom2}
\docType{data}

\title{
Association between the gut microbiome and Clostridium difficile infections
}

\description{
Real data on Clostridium difficile infections
}

\usage{data(biom2)}
\format{
phyloseq
}

\references{
McMurdie PJ and Holmes S. phyloseq: An R package for reproducible interactive analysis and graphics of microbiome census data. PLoS ONE. 2013;8(4):e61217

Schubert AM, et al. Microbiome data distinguish patients with Clostridium difficile infection and Non-C. difficile-associated diarrhea from healthy controls. mBio 2014, 5(3):e01021–14.
}

\examples{
library(phyloseq)
data(biom2)
}
