\name{phy.parameters}
\alias{phy.parameters}
\docType{data}

\title{
Estimated parameters of Dirichlet-multinomial distribution
}

\description{
Real data on colorectal cancer
}

\usage{data(phy.parameters)}
\format{
phyloseq
}

\references{
Koh H and Zhao N. A powerful microbial group association test based on the higher criticism analysis for sparse microbial association signals. Microbiome 2020;8(1):63.

Wu C, et al. An adaptive association test for microbiome data. Genome Med 2016;8(1):56.
}

\examples{
data(phy.parameters)
}
